chore: Rename user api file
