+++ fatog-fe/src/pages/users/UserForm.jsx
@@ -6,8 +6,8 @@ import { useForm } from 'react-hook-form';
 import { Stack, HStack, Flex, Box, Icon, IconButton, Spinner, Button, Heading, FormLabel, Text } from '@chakra-ui/react';
 import Breadcrumb from '../../components/Breadcrumb';
 import SelectElement from '../../components/form/SelectElement';
-import { createUser, updateUser } from '../../api/user';
-import { getUser } from '../../api/user';
+import { createUser, updateUser } from '../../api/users';
+import { getUser } from '../../api/users';
 import { requireAuth } from '../../hooks/useAuth';
 import { useToastHook } from '../../hooks/useToast';
 import { BiError } from "react-icons/bi";
